+++ src/Accommodation/AccommodationDialogContentEdit.tsx
@@ -2,7 +2,8 @@ import { Box, Dialog, Spinner } from '@radix-ui/themes';
 import { DateTime } from 'luxon';
 import { useCallback } from 'react';
 import type { DialogContentProps } from '../Dialog/DialogRoute';
-import { type TripSliceAccommodation, useTrip } from '../Trip/store';
+import { useTrip } from '../Trip/hooks';
+import type { TripSliceAccommodation } from '../Trip/store';
 import { AccommodationDialogMode } from './AccommodationDialogMode';
 import { AccommodationForm } from './AccommodationForm';
 import { AccommodationFormMode } from './AccommodationFormMode';
