+++ src/Activity/ActivityDialogContentEdit.tsx
@@ -2,7 +2,8 @@ import { Box, Dialog, Spinner } from '@radix-ui/themes';
 import { DateTime } from 'luxon';
 import { useCallback } from 'react';
 import type { DialogContentProps } from '../Dialog/DialogRoute';
-import { type TripSliceActivity, useTrip } from '../Trip/store';
+import { useTrip } from '../Trip/hooks';
+import type { TripSliceActivity } from '../Trip/store';
 import { ActivityDialogMode } from './ActivityDialogMode';
 import { ActivityForm } from './ActivityForm';
 import { ActivityFormMode } from './ActivityFormMode';
