+++ src/Activity/ActivityDialogContentView.tsx
@@ -14,7 +14,8 @@ import { useParseTextIntoNodes } from '../common/text/parseTextIntoNodes';
 import type { DialogContentProps } from '../Dialog/DialogRoute';
 import { CommonCommentDialogMaxWidth } from '../Dialog/ui';
 import { useBoundStore } from '../data/store';
-import { type TripSliceActivity, useTrip } from '../Trip/store';
+import { useTrip } from '../Trip/hooks';
+import type { TripSliceActivity } from '../Trip/store';
 import s from './Activity.module.css';
 import { ActivityMap } from './ActivityDialogMap';
 import { ActivityDialogMode } from './ActivityDialogMode';
