+++ src/Expense/ExpenseRow.tsx
@@ -3,7 +3,8 @@ import { Button, Flex, Popover, Table, Text } from '@radix-ui/themes';
 import { DateTime } from 'luxon';
 import { useState } from 'react';
 import { useBoundStore } from '../data/store';
-import { type TripSliceExpense, useTrip } from '../Trip/store';
+import { useTrip } from '../Trip/hooks';
+import type { TripSliceExpense } from '../Trip/store';
 import { dangerToken } from '../ui';
 import { dbDeleteExpense } from './db';
 import { ExpenseInlineForm } from './ExpenseInlineForm';
