+++ src/Macroplan/MacroplanDialogContentEdit.tsx
@@ -2,7 +2,8 @@ import { Box, Dialog, Spinner } from '@radix-ui/themes';
 import { DateTime } from 'luxon';
 import { useCallback } from 'react';
 import type { DialogContentProps } from '../Dialog/DialogRoute';
-import { type TripSliceMacroplan, useTrip } from '../Trip/store';
+import { useTrip } from '../Trip/hooks';
+import type { TripSliceMacroplan } from '../Trip/store';
 import { MacroplanDialogMode } from './MacroplanDialogMode';
 import { MacroplanForm } from './MacroplanForm';
 import { MacroplanFormMode } from './MacroplanFormMode';
