+++ src/Macroplan/MacroplanDialogContentView.tsx
@@ -10,7 +10,8 @@ import { DateTime } from 'luxon';
 import { useCallback } from 'react';
 import { useParseTextIntoNodes } from '../common/text/parseTextIntoNodes';
 import type { DialogContentProps } from '../Dialog/DialogRoute';
-import { type TripSliceMacroplan, useTrip } from '../Trip/store';
+import { useTrip } from '../Trip/hooks';
+import type { TripSliceMacroplan } from '../Trip/store';
 import s from './Macroplan.module.css';
 import { MacroplanDialogMode } from './MacroplanDialogMode';
