+++ src/Trip/TripEditDialog.tsx
@@ -2,7 +2,8 @@ import { Box, Dialog } from '@radix-ui/themes';
 import { DateTime } from 'luxon';
 import { CommonDialogMaxWidth } from '../Dialog/ui';
 import { useBoundStore } from '../data/store';
-import { type TripSliceTrip, useTripActivities } from './store';
+import { useTripActivities } from './hooks';
+import type { TripSliceTrip } from './store';
 import { TripForm } from './TripForm';
 import { TripFormMode } from './TripFormMode';
 import { formatToDateInput } from './time';
